systems: Move spring_mass_system to a test utility
This allows us to remove the awkward and misleading drake/system/plants
directory.

This is a breaking change for anyone who was using this plant via the
precompiled binaries.
