+++ systems/analysis/BUILD.bazel
@@ -507,7 +507,7 @@ drake_cc_googletest(
         "//common/test_utilities:expect_throws_message",
         "//systems/analysis/test_utilities:implicit_integrator_test",
         "//systems/analysis/test_utilities:quadratic_scalar_system",
-        "//systems/plants/spring_mass_system",
+        "//systems/analysis/test_utilities:spring_mass_system",
     ],
 )
 
@@ -516,7 +516,7 @@ drake_cc_googletest(
     deps = [
         ":implicit_integrator",
         "//common/test_utilities:expect_no_throw",
-        "//systems/plants/spring_mass_system",
+        "//systems/analysis/test_utilities:spring_mass_system",
     ],
 )
 
@@ -525,7 +525,7 @@ drake_cc_googletest(
     deps = [
         ":integrator_base",
         "//common/test_utilities:expect_throws_message",
-        "//systems/plants/spring_mass_system",
+        "//systems/analysis/test_utilities:spring_mass_system",
     ],
 )
 
@@ -542,7 +542,7 @@ drake_cc_googletest(
         "//systems/analysis/test_utilities:implicit_integrator_test",
         "//systems/analysis/test_utilities:linear_scalar_system",
         "//systems/analysis/test_utilities:quadratic_scalar_system",
-        "//systems/plants/spring_mass_system",
+        "//systems/analysis/test_utilities:spring_mass_system",
     ],
 )
