+++ systems/analysis/test/implicit_integrator_test.cc
@@ -2,7 +2,7 @@
 
 #include <gtest/gtest.h>
 
-#include "drake/systems/plants/spring_mass_system/spring_mass_system.h"
+#include "drake/systems/analysis/test_utilities/spring_mass_system.h"
 
 using Eigen::VectorXd;
