+++ systems/analysis/test/simulator_test.cc
@@ -21,10 +21,10 @@
 #include "drake/systems/analysis/test_utilities/controlled_spring_mass_system.h"
 #include "drake/systems/analysis/test_utilities/logistic_system.h"
 #include "drake/systems/analysis/test_utilities/my_spring_mass_system.h"
+#include "drake/systems/analysis/test_utilities/spring_mass_system.h"
 #include "drake/systems/analysis/test_utilities/stateless_system.h"
 #include "drake/systems/framework/diagram_builder.h"
 #include "drake/systems/framework/event.h"
-#include "drake/systems/plants/spring_mass_system/spring_mass_system.h"
 #include "drake/systems/primitives/constant_vector_source.h"
 #include "drake/systems/primitives/integrator.h"
 #include "drake/systems/primitives/signal_logger.h"
