+++ systems/analysis/test_utilities/controlled_spring_mass_system.h
@@ -4,9 +4,9 @@
 
 #include "drake/common/default_scalars.h"
 #include "drake/common/drake_copyable.h"
+#include "drake/systems/analysis/test_utilities/spring_mass_system.h"
 #include "drake/systems/controllers/pid_controller.h"
 #include "drake/systems/framework/diagram.h"
-#include "drake/systems/plants/spring_mass_system/spring_mass_system.h"
 #include "drake/systems/primitives/constant_vector_source.h"
 
 namespace drake {
