+++ systems/analysis/test_utilities/spring_mass_damper_system.h
@@ -3,7 +3,7 @@
 #include <limits>
 #include <utility>
 
-#include "drake/systems/plants/spring_mass_system/spring_mass_system.h"
+#include "drake/systems/analysis/test_utilities/spring_mass_system.h"
 
 // WARNING WARNING WARNING
 // This test is currently used only as a stiff system test for implicit
