+++ systems/analysis/test_utilities/spring_mass_system.cc
@@ -1,4 +1,4 @@
-#include "drake/systems/plants/spring_mass_system/spring_mass_system.h"
+#include "drake/systems/analysis/test_utilities/spring_mass_system.h"
 
 #include <utility>
