--- systems/plants/spring_mass_system/BUILD.bazel
@@ -1,43 +1,0 @@
-# -*- python -*-
-
-load(
-    "@drake//tools/skylark:drake_cc.bzl",
-    "drake_cc_googletest",
-    "drake_cc_library",
-    "drake_cc_package_library",
-)
-load("//tools/lint:lint.bzl", "add_lint_tests")
-
-package(default_visibility = ["//visibility:public"])
-
-drake_cc_package_library(
-    name = "spring_mass_system",
-    visibility = ["//visibility:public"],
-    deps = [
-        ":everything",
-    ],
-)
-
-drake_cc_library(
-    name = "everything",
-    srcs = ["spring_mass_system.cc"],
-    hdrs = ["spring_mass_system.h"],
-    visibility = ["//visibility:private"],
-    deps = [
-        "//systems/framework",
-    ],
-)
-
-# === test/ ===
-
-drake_cc_googletest(
-    name = "spring_mass_system_test",
-    size = "medium",
-    deps = [
-        ":spring_mass_system",
-        "//common/test_utilities",
-        "//systems/framework/test_utilities",
-    ],
-)
-
-add_lint_tests()
